Time: 3 ms (100%), Space: 43.8 MB (62.37%) - LeetHub

+++ 0201-bitwise-and-of-numbers-range/0201-bitwise-and-of-numbers-range.java
@@ -1,6 +1,6 @@
 class Solution {
     public int rangeBitwiseAnd(int left, int right) {
-        int i=0;
+        short i=0;
         while(left != right){
             left >>= 1;
             right >>= 1;
